README: added some describtion
Commiting it in it's imperfect form, because moving on forward seem to be better than holding it off and still be thinking about it for a while after.

There. It is done. About this part I don't have to worry.
I will return and pick up where I left, so it ends up the same, except the thoughts aren't bothering me.

Calling it a day, now.